Реализован пакет ресурсов, сделал пробный запуск тестов.

+++ src/main/java/requests/skelethon/interfaces/CrudEndpointInterface.java
@@ -4,7 +4,7 @@ import models.BaseModel;
 
 public interface CrudEndpointInterface {
     Object post(BaseModel model);
-    Object get(int id);
-    Object update(int id, BaseModel model);
-    Object delete(int id);
+    Object get(long id);
+    Object update(long id, BaseModel model);
+    Object delete(long id);
 }
